requirements.txt
SpeechRecognition==3.8.1
python-telegram-bot==20.0a4
pandas==1.4.3